+++ src/arrayBufferToBase64.ts
@@ -1,0 +1,11 @@
+/**
+ * String, number to base64
+ *
+ * @param input - the string to be encoded
+ * @returns - the BASE64 encoding
+ */
+function arrayBufferToBase64(input: string) {
+	//
+}
+
+export default arrayBufferToBase64