+++ src/base64ToBlob.ts
@@ -1,0 +1,11 @@
+/**
+ * String, number to base64
+ *
+ * @param input - the string to be encoded
+ * @returns - the BASE64 encoding
+ */
+function base64ToBlob(input: string) {
+	//
+}
+
+export default base64ToBlob